+++ matlab/CHANGES.md
@@ -1,5 +1,8 @@
 # Cepton SDK MATLAB Release Notes
 
+### Version 1.8 2018-10-08
+* Update SDK.
+
 ### Version 1.7 2018-10-02
 * Add `cepton_sdk_matlab_clear_cache` function.
 * Lots of minor bugfixes.